+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Head
 
+* Added: the `indentation` rule now checks indentation of multi-line at-rule params, unless there's the `except` option of `param`.
 * Fixed: the `color-no-invalid-hex` rule no longer flags hashes in `url()` arguments.
 * Added: support for end-of-line comments in `selector-list-comma-newline-after`.
 * Added: protection against `#${sass-interpolation}` in rules checking for hex colors.